add: adding request validation codec
